requirements.txt
numpy
matplotlib
gym
gym[atari]
torch
torchvision
tensorflow
sklearn
pyyaml
pyparsing
tqdm